servers added now show up in main page

+++ varnishinterface/get_varnish_servers.php
@@ -2,8 +2,13 @@
 	session_start();
 
 	include_once( 'config.php' );
-?>
 
+	$showManagement=false;
+	if(isset($_GET['manage'])) {
+		$showManagement=true;
+	}
+	if ($showManagement) {
+?>
 <table class="table table-hover">
 <tbody>
 <tr>
@@ -11,15 +16,30 @@
 <td>Delete server</td>
 </tr>
 <?php
+	}
 	$i=0;
 	foreach (glob("$varnish_secret_path_prefix*") as $filename)	{
-		echo "<tr>";
-		echo "<td id=ipAddr$i>" . substr($filename, strlen($varnish_secret_path_prefix)) .  "</td>";
-		echo "<td><a href='javascript:deleteVarnish($i);'><span class='glyphicon glyphicon-trash'></span></a></td>";
-		echo "</tr>";
+		if( $showManagement ) {
+			echo "<tr>";
+			echo "<td id=ipAddr$i>" . substr($filename, strlen($varnish_secret_path_prefix)) .  "</td>";
+			echo "<td><a href='javascript:deleteVarnish($i);'><span class='glyphicon glyphicon-trash'></span></a></td>";
+			echo "</tr>";
+		}
+		else {
+?>
+			<a class="button" href="#" onclick="loadAjax('get_varnish_stat.php?ip='+getElementById('ip<?php echo $i;?>').innerHTML+'&port='+getElementById('port').value,'vsdata');" ><span id="ip<?php echo $i;?>">
+<?php
+			echo substr($filename, strlen($varnish_secret_path_prefix));	
+			echo "</span></a>";
+		}
 		$i++;
 	}
+
+	if( $showManagement ) {
 ?>
 </tbody>
 </table>
 </div>
+<?php
+		}
+?>
